requirements.txt
# For PyTorch with CUDA: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu126

opencv-python
numpy
# torch
# torchvision
# torchaudio
transformers
gTTS
pygame
pillow
requests
google-generativeai
pandas
ultralytics
timm